Gorlex DLC 2: Winter Coats! (#3400)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Wow, didn't expect me to be back so early huh?

The Main Gorlex Splinters now have Winter Coats!

![image](https://github.com/user-attachments/assets/10926115-0384-4592-b4bf-8d2bab7fe84a)

![image](https://github.com/user-attachments/assets/f4ae8f99-98f2-42f0-8c52-326a6a0c0308)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
I never want to hear the excuse of "It gives cold protection" for
wearing a hermit uniform in an NGR ship ever again.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: PositiveEntropy
add: Gorlex Splinters now have winter coats!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->