minimum recoil for ballistics (#3330)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
gives a very small kick to all ballistics, not enough to affect your aim
but makes them feel alot better to use.
atomization of #3094
currently debating what the value should be..

0

https://github.com/user-attachments/assets/1630a089-bf9f-4ec5-8853-c2728f420f4e

.1

https://github.com/user-attachments/assets/2064b49b-37f5-4860-b273-1f28274724f7

.175

https://github.com/user-attachments/assets/6baba2ad-8100-403e-b6c4-d4e958ad21d6

.25

https://github.com/user-attachments/assets/ca3bdf37-bf30-493e-bb24-ae452cbc6481

a bit of an extreme case being a fast firing auto but it happens to be a
gun with a recoil of 0 rn so it will be affected by the change

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
without it, a surprising amount of ballistics get 0 recoil just but
wielding them which feels really jarring.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Ballistics now have a minimum recoil, not enough to mess up your
shot!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->