Waste Planet Turf Resprite! (#3473)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Does what it says on the tin, redoing all the turf tiles related to the
waste planet!


![image](https://github.com/user-attachments/assets/a44e988d-9fbc-4326-aaba-50d0310293a0)

![image](https://github.com/user-attachments/assets/bdb122c8-ac3b-40a5-a98b-b3caa1caa5d5)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The previous one was so noisy and grody, it was high time they had a
visual update!
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: PositiveEntropy
imageadd: Waste Planet Turfs Now Look Much More Refined!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->